requirements.txt
discord.py
python-dotenv
tinydb
#markdown
#markupsafe
#flask
